% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstDIF-package.R
\docType{package}
\name{mstDIF-package}
\alias{mstDIF-package}
\title{mstDIF: A Collection of Statistical Tests for DIF Detection in Multistage Tests}
\description{
A Collection of Statistical Tests for the Detection of Differential
Item Functioning (DIF) in Multistage Tests. Methods entail logistic regression,
mstSIB, and various score-based tests.
}
\author{
\strong{Maintainer}: Rudolf Debelak \email{rudolf.debelak@gmail.com}

Authors:
\itemize{
  \item Dries Debeer \email{debeer.dries@gmail.com}
}

Other contributors:
\itemize{
  \item Sebastian Appelbaum [contributor]
  \item Mark J. Gierl [contributor]
}

}
\keyword{internal}
